+++ NavigationController/NavigationController/LibraryAPI.h
@@ -15,8 +15,8 @@
 
 
 - (NSArray*)getAlbums;
-- (void)addAlbum:(Album*)album atIndex:(int)index;
-- (void)deleteAlbumAtIndex:(int)index;
+- (void)addAlbum:(Album*)album atIndex:(NSInteger)index;
+- (void)deleteAlbumAtIndex:(NSInteger)index;
 
 - (void)saveAlbums;
